Desktop/SamiraParser/requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
nltk==3.8.1
python-docx==1.0.1
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Bootstrap-Flask==2.3.0 